requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
wordcloud
emoji
urlextract
fpdf2
nltk
textblob
python-dateutil
